+++ modules/kubernetes/aks-core/variables.tf
@@ -277,6 +277,47 @@ variable "datadog_config" {
   }
 }
 
+variable "grafana_agent_enabled" {
+  description = "Should Grafana-Agent be enabled"
+  type        = bool
+  default     = false
+}
+
+variable "grafana_agent_config" {
+  description = "The Grafan-Agent configuration"
+  sensitive   = true
+  type = object({
+    remote_write_urls = object({
+      metrics = string
+      logs    = string
+      traces  = string
+    })
+    credentials = object({
+      metrics_username = string
+      metrics_password = string
+      logs_username    = string
+      logs_password    = string
+      traces_username  = string
+      traces_password  = string
+    })
+  })
+  default = {
+    remote_write_urls = {
+      metrics = ""
+      logs    = ""
+      traces  = ""
+    }
+    credentials = {
+      metrics_username = ""
+      metrics_password = ""
+      logs_username    = ""
+      logs_password    = ""
+      traces_username  = ""
+      traces_password  = ""
+    }
+  }
+}
+
 variable "falco_enabled" {
   description = "Should Falco be enabled"
   type        = bool