+++ modules/kubernetes/prometheus/variables.tf
@@ -134,6 +134,12 @@ variable "azad_kube_proxy_enabled" {
   default     = false
 }
 
+variable "grafana_agent_enabled" {
+  description = "Should grafana-agent be enabled"
+  type        = bool
+  default     = false
+}
+
 # This should be removed, https://github.com/kubernetes/kube-state-metrics/issues/1631
 variable "kube_state_metrics_namespaces" {
   description = "Comma-separated list of namespaces to be enabled. To get metrics from all namespaces use ''"